update and fix feats

+++ app/Http/Controllers/ResidenceBuildingController.php
@@ -115,6 +115,9 @@ class ResidenceBuildingController extends Controller
      */
     public function destroy( Residence $residence, Building $building)
     {
+        if ($building->floors()->exists()) {
+            return redirect()->route('residences.buildings.index', $residence)->withErrors(['message' => 'Impossible to delete a building that contains floors.']);
+        }
         $building->delete();
 
         return redirect()->route('residences.buildings.index', $residence)->with('success', 'Le Bâtiment à bien été supprimée');
